on_message if-statement, slash-commands, question
There was indeed a bug in the on_message if-statement, and it has been squashed!

Slash commands have been added to add and remove channels to the bot as you please, which are written and read to/from a new JSON file.

Questions are now also added via a JSON file, which is a manual process.

May add a slash-command to use Chat-GPT...